chat forum, manage article

+++ routes/web.php
@@ -58,10 +58,12 @@ Route::get('/posts/{post:slug}', [PostController::class, 'show']);
 
 Route::get('home/{post:slug}', [PostController::class, 'show']);
 
-Route::get('article', [PostController::class, 'article'])->name('article');
-Route::post('article/edit', [PostController::class, 'edit'])->name('article.edit');
-Route::delete('/data/{id}', [PostController::class, 'delete'])->name('deleteArc');
-Route::post('/article/editpost', [PostController::class, 'editpost'])->name('article.editpost');
+Route::get('article', [PostController::class, 'article'])->name('article')->middleware('admin');
+Route::get('article/create', [PostController::class, 'create'])->name('article.create')->middleware('admin');
+Route::post('article/edit', [PostController::class, 'edit'])->name('article.edit')->middleware('admin');
+Route::delete('/data/{id}', [PostController::class, 'delete'])->name('deleteArc')->middleware('admin');
+Route::post('/article/editpost', [PostController::class, 'editpost'])->name('article.editpost')->middleware('admin');
+Route::post('/article/createpost', [PostController::class, 'createpost'])->name('article.createpost')->middleware('admin');
 
 
 Route::get('/login', [LoginController::class, 'index'])->name('login')->middleware('guest');
@@ -78,23 +80,23 @@ Route::get('/dashboard', function(){
 
 })->middleware('auth');
 
-Route::get('/forum', [TopicController::class, 'index'])->name('topics.index');
-Route::get('/forum/topics/create', [TopicController::class, 'create'])->name('topics.create');
-Route::post('/forum/topics', [TopicController::class, 'store'])->name('topics.store');
+Route::get('/forum', [TopicController::class, 'index'])->name('topics.index')->middleware('loginonly');
+Route::get('/forum/topics/create', [TopicController::class, 'create'])->name('topics.create')->middleware('loginonly');
+Route::post('/forum/topics', [TopicController::class, 'store'])->name('topics.store')->middleware('loginonly');
 // Route::get('/forum/topics/{topic}', [TopicController::class, 'show'])->name('topics.show');
-Route::post('/forum/topic/detail', [TopicController::class, 'show'])->name('topics.show');
-Route::post('/forum/topics/{topic}/replies', [ReplyController::class, 'store'])->name('replies.store');
+Route::post('/forum/topic/detail', [TopicController::class, 'show'])->name('topics.show')->middleware('loginonly');
+Route::post('/forum/topics/{topic}/replies', [ReplyController::class, 'store'])->name('replies.store')->middleware('loginonly');
 
-Route::get('/getReplies', [ReplyController::class, 'getReplies'])->name('getReplies');
+Route::get('/getReplies', [ReplyController::class, 'getReplies'])->name('getReplies')->middleware('loginonly');
 
 
-Route::get('/moods', [MoodController::class, 'index'])->name('moods.index');
-Route::get('/moods/create', [MoodController::class, 'create'])->name('moods.create');
-Route::get('/moods/activity', [MoodController::class, 'activity'])->name('moods.activity');
-Route::get('/moods/result', [MoodController::class, 'result'])->name('moods.result');
-Route::get('/moods/history', [MoodController::class, 'history'])->name('moods.history');
-Route::post('/moods', [MoodController::class, 'store'])->name('moods.store');
-Route::post('/moodstodo', [MoodController::class, 'storetodo'])->name('moods.storetodo');
+Route::get('/moods', [MoodController::class, 'index'])->name('moods.index')->middleware('loginonly');
+Route::get('/moods/create', [MoodController::class, 'create'])->name('moods.create')->middleware('loginonly');
+Route::get('/moods/activity', [MoodController::class, 'activity'])->name('moods.activity')->middleware('loginonly');
+Route::get('/moods/result', [MoodController::class, 'result'])->name('moods.result')->middleware('loginonly');
+Route::get('/moods/history', [MoodController::class, 'history'])->name('moods.history')->middleware('loginonly');
+Route::post('/moods', [MoodController::class, 'store'])->name('moods.store')->middleware('loginonly');
+Route::post('/moodstodo', [MoodController::class, 'storetodo'])->name('moods.storetodo')->middleware('loginonly');
 
 
 
@@ -112,6 +114,6 @@ Route::get('/doctors', [DoctorController::class, 'index'])->name('doctors.index'
 Route::get('dashboard/doctors/create', [DoctorController::class, 'create'])->name('doctors.create');
 Route::post('/doctors', [DoctorController::class, 'store'])->name('doctors.store');
 
-Route::get('/profile/edit', [ProfileController::class, 'edit'])->name('profile.edit');
-Route::post('/profile/update', [ProfileController::class, 'update'])->name('profile.update')->middleware('auth');
-Route::get('/places/find-nearest-psychiatrists', 'App\Http\Controllers\PlacesController@findNearestPsychiatrists');
+Route::get('/profile/edit', [ProfileController::class, 'edit'])->name('profile.edit')->middleware('loginonly');
+Route::post('/profile/update', [ProfileController::class, 'update'])->name('profile.update')->middleware('auth')->middleware('loginonly');
+Route::get('/places/find-nearest-psychiatrists', 'App\Http\Controllers\PlacesController@findNearestPsychiatrists')->middleware('loginonly');
